refactor(ui): centralize UTM persistence into shared package
Moved utm.ts and utm-persistence.tsx into @prisma-docs/ui so all
three apps (site, blog, docs) share a single implementation.

Each app's local utm-persistence.tsx is now a thin wrapper that
passes its config (storageKey, basePath or proxiedPaths) to the
shared component. The local utm.ts files are deleted — consumers
import directly from @prisma-docs/ui/lib/utm.
